Finish through_class function
through_class function has been updated, with changes to both its
parameters and its output. It no longer receives a list but instead
takes in a class name and an href that class page. It first creates a
Class object with the name. It will then call grab_classes to get the text and grades from every review and put them in the reviews list of the created class object.
This further containerization of reviews from professors to classes will
allow for easier department based analysis