Add every_class()/ grab_classes->grab_reviews
Change grab_classes to grab_reviews to avoid confusion

Implementation of every_class function. This function is called by
all_classes to allow for multi page navigability. every_class takes over
the job of loading all of the webpages and then return a href if there
is another page of classes, a 0 if not, very similar to how grab_reviews
works.

Every_class also calls through_class with the name and href of each
class and appends the result to the professors class list. Therefore,
the output of calling all_classes will be a professor object whose list
of classes is full of list objects, which are subsequently full of
review objects that contain information about each review