Problem/Solution for insufficient prof database
Turns out, there there is no consecutive page system on Bruinwak that
contains every professor. To get around this, I am going to find the
hrefs for each department and then run the existing code on that.

One problem that I've run into is that the list of departments isn't
accessible through Bruinwalk's html. However, the url change between
deparments is just one number. So I am implementing the function
get_department() to create potential department urs and then check if
they are valid, adding the url to a list if so.